feat(orderbook): implement std::map-based OrderBook with price levels
- add OrderBook skeleton using std::map
- implement bid/ask insertion
- implement best bid and best ask lookup
- fix reference vs copy mutation bug
- fix order overwrite by appending to price level
- add unit tests for OrderBook
- document Chapter 2 design decisions and trade-offs

+++ include/chimera/price_level.hpp
@@ -1,9 +1,10 @@
 #pragma once
 #include <list>
 #include "order.hpp"
+#include <cstdint>
 
 struct PriceLevel
 {
-    int price;
+    int64_t price;
     std::list<Order> orders;
 };
